% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfgsi.R
\name{bfgsi}
\alias{bfgsi}
\title{BFGS update}
\usage{
bfgsi(H0, dg, dx)
}
\arguments{
\item{H0}{Previous inverse Hessian}

\item{dg}{previous change in gradient}

\item{dx}{previous change in location}
}
\description{
Updates inverse Hessian based on one iteration of a gradient-based optimization routine.
}
