% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numgrad.R
\name{numgrad}
\alias{numgrad}
\title{Numerical gradient}
\usage{
numgrad(fcn, x, ...)
}
\arguments{
\item{fcn}{The function to be differentiated (can be vector-valued),}

\item{x}{The point around which to differentiate.}

\item{...}{Other arguments to \code{fcn}.}
}
\value{
\item{g}{numerical gradient}
\item{badg}{Is there a numerical problem with the gradient?}
}
\description{
Simple delta f / delta x numerical gradient
}
\details{
You might need to change the \code{delta} value, depending on your
application. \code{badg} probably is true only when \code{fcn} returns a huge
value to signal \code{x} outside its domain of definition, or NaN's.
}
