% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numHess.R
\name{numHess}
\alias{numHess}
\title{numHess}
\usage{
numHess(fcn, x, ...)
}
\arguments{
\item{fcn}{The function to be differentiated.}

\item{x}{The point at which the Hessian is to be calculated.}
}
\description{
Numerical Hessian.  This is a naive calculation, just
applying \code{numgrad()}'s simple forward difference estimate
twice.
}
