% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(x)
}
\arguments{
\item{`x`}{a vector of arbitrary dimension n}
}
\value{
Minus the height of the volcano at \code{x}
}
\description{
A "tilted volcano" objective function for testing optimizers
}
\details{
Gradient based routines may tend to go to the volcano rim, then
only very slowly crawl along it to the peak.
}
