% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numHess.R
\name{numHess}
\alias{numHess}
\title{Numerical Hessian}
\usage{
numHess(fcn, x, ...)
}
\arguments{
\item{fcn}{The function to differentiate.}

\item{x}{The point at which to differentiate.}

\item{...}{Additional arguments for \code{fcn}.}
}
\value{
Numerical second derivative matrix
}
\description{
Naive numerical Hessian.  Just applies numgrad twice.
}
